django-redis-test/mysite/requirements.txt
Django==3.1.1
django-debug-toolbar==3.1.1
django-redis==4.12.1
